requirements.txt
streamlit
openai
getllm
openai-whisper
torch
mysql-connector-python